Added new test cases
